requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt -
#
appdirs==1.4.4            # via black
attrs==19.3.0             # via black, pytest
beautifulsoup4==4.9.1     # via -r -
black==19.10b0            # via -r -
click==7.1.2              # via -r -, black, pip-tools
colorful==0.5.4           # via prettyprinter
icdiff==1.9.1             # via -r -
importlib-metadata==1.7.0  # via pluggy, pytest
lark-parser==0.9.0        # via -r -
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.2.1          # via -r -
pluggy==0.13.1            # via pytest
prettyprinter==0.18.0     # via -r -
py==1.9.0                 # via pytest
pygments==2.6.1           # via prettyprinter
pyparsing==2.4.7          # via packaging
pytest==5.4.3             # via -r -
regex==2020.7.14          # via black
six==1.15.0               # via packaging, pip-tools
soupsieve==2.0.1          # via beautifulsoup4
toml==0.10.1              # via black
typed-ast==1.4.1          # via black
wcwidth==0.2.5            # via pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
